+++ navigation/lqr/nodes/lqr_node.py
@@ -10,10 +10,64 @@ import numpy.typing as npt
 import matplotlib.pyplot as plt
 import tf2_ros
 from nav_msgs.msg import Path
-from lqr import prepTrack, Track, SolverMatrices, setupMatrices, optimizeMinCurve, createRaceLine, pathToNumpy, addWidth
+from lqr import prepTrack, Track, SolverMatrices, setupMatrices, optimizeMinCurve, createRaceLine
 
 TRACK_WIDTH = rospy.get_param("/navigation/lqr/handler/track_width")
 SAFETY_MARGIN = rospy.get_param("/navigation/lqr/handler/safety_margin")
+
+
+def addWidth(trackNoWidth: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
+
+    """
+    Adds width of track to track array if only the x and y coordinates of the track are given
+
+    Parameters
+    ----------
+    trackNoWidth: np.ndarray, shape=(n, 2)
+        Track array with only x and y coordinates
+
+    Returns
+    -------
+    wideTrack: np.ndarray, shape=(n, 4)
+        Track array with x, y, left width and right width
+    """
+
+    widthCol = np.ones(int(trackNoWidth.shape[0])) * (TRACK_WIDTH - SAFETY_MARGIN)
+    trackNoWidth = np.stack(
+        (
+            trackNoWidth[:, 0],
+            trackNoWidth[:, 1],
+            widthCol / 2,
+            widthCol / 2,
+        ),
+        axis=1,
+    )
+    wideTrack = np.reshape(trackNoWidth, (-1, 4))
+    return wideTrack
+
+
+def pathToNumpy(path: Path) -> npt.NDArray[np.float64]:
+    """
+    Converts a path message to a numpy array
+    and transforms the coordinates to the world frame
+
+    Parameters
+    ----------
+    path: Path
+        Path message
+
+    Returns
+    -------
+    pathArray: np.ndarray, shape=(n, 2)
+        Numpy array with x and y coordinates of the path
+    """
+    tfBuffer = tf2_ros.Buffer()
+    pathArray = np.zeros((len(path.poses), 2))
+    for index, pose in enumerate(path.poses):
+        globalPose = tfBuffer.transform(pose, "world", rospy.Duration(1))
+        pathArray[index, 0] = globalPose.pose.position.x
+        pathArray[index, 1] = globalPose.pose.position.y
+    return pathArray
 
 
 def main() -> None:
